API-45: Extract cursor codec into api/util.
